pyproject.toml
[tool.poetry]
name = "commit-task"
version = "0.1.0"
description = ""
authors = ["vliubko <neznam.ua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
requests = "^2.32.3"
pymongo = "^4.11"
python-dotenv = "^1.0.1"
python-json-logger = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
